function_app/requirements.txt
azure-functions
azure-identity
azure-data-tables
azure-core

